Increment min date to avoid timezone issues
The date given is treated as date in local timezone, so when internally converted to UTC it goes before epoch start time and messes things up (leading to overflow error). So set it to 2nd Jan 1970, which makes no practical difference to the rest of the code but avoids this issue.